refactor: rename project from pypi-server to kingpi
Rename the Python package from pypi_server to kingpi and update all
references across the codebase. The old name was confusing since this
project is a wrapper/proxy around PyPI, not PyPI itself.

- Rename src/pypi_server/ → src/kingpi/
- Update all imports, env prefix (KINGPI_), db default, FastAPI title
- Keep PyPIClient/pypi_client.py as-is since it's a client *for* PyPI
