+++ src/kingpi/app.py
@@ -3,7 +3,7 @@ from fastapi import FastAPI
 
 def create_app() -> FastAPI:
     app = FastAPI(
-        title="PyPI Server",
+        title="KingPi",
         description="Lightweight PyPI package analytics server",
         version="0.1.0",
     )
