+++ src/kingpi/schemas/event.py
